requirements.txt
watchdog
regex
requests
pyenchant